+++ src/components/AtmosWorkflow/constants.ts
@@ -1,0 +1,4 @@
+// constants.ts
+
+export const CLOUDPOSSE_DOCS_URL = 'https://raw.githubusercontent.com/cloudposse/docs/master/';
+export const WORKFLOWS_DIRECTORY_PATH = 'examples/snippets/stacks/workflows/';
