+++ src/components/AtmosWorkflow/types.ts
@@ -1,0 +1,6 @@
+// types.ts
+
+export interface WorkflowStep {
+  type: 'command' | 'title';
+  content: string;
+}
